build: implement MVP architecture for Lotto feature with LottoPresenter and LottoContract
